updated commands
